+++ src/base_screen.lua
@@ -9,4 +9,7 @@ end
 function class:on_key_pressed(key, scancode, is_repeat)
 end
 
+function class:on_mouse_pressed(x, y, button, is_touch, presses)
+end
+
 return class